introduces a JndiConfigurator that is run before SlickProvider creation
