+++ persistence-jdbc/core/src/test/scala/com/lightbend/lagom/internal/persistence/jdbc/SlickDbTestProvider.scala
@@ -30,4 +30,5 @@ object SlickDbTestProvider {
 
     new InitialContext().rebind("DefaultDB", slickDb)
   }
+
 }
